Test fix Changed `hidden.py` and other files to make it work with the latest version of PyTorch.

+++ model/hidden.py
@@ -139,9 +139,9 @@ class Hidden:
         self.encoder_decoder.eval()
         self.discriminator.eval()
         with torch.no_grad():
-            d_target_label_cover = torch.full((batch_size, 1), self.cover_label, device=self.device)
-            d_target_label_encoded = torch.full((batch_size, 1), self.encoded_label, device=self.device)
-            g_target_label_encoded = torch.full((batch_size, 1), self.cover_label, device=self.device)
+            d_target_label_cover = torch.full((batch_size, 1), self.cover_label, device=self.device).float()
+            d_target_label_encoded = torch.full((batch_size, 1), self.encoded_label, device=self.device).float()
+            g_target_label_encoded = torch.full((batch_size, 1), self.cover_label, device=self.device).float()
 
             d_on_cover = self.discriminator(images)
             d_loss_on_cover = self.bce_with_logits_loss(d_on_cover, d_target_label_cover)
